# src/routes/valor.js
"use strict";
var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
    return new (P || (P = Promise))(function (resolve, reject) {
        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
        step((generator = generator.apply(thisArg, _arguments || [])).next());
    });
};
var __generator = (this && this.__generator) || function (thisArg, body) {
    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
    function verb(n) { return function (v) { return step([n, v]); }; }
    function step(op) {
        if (f) throw new TypeError("Generator is already executing.");
        while (g && (g = 0, op[0] && (_ = 0)), _) try {
            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
            if (y = 0, t) op = [op[0] & 2, t.value];
            switch (op[0]) {
                case 0: case 1: t = op; break;
                case 4: _.label++; return { value: op[1], done: false };
                case 5: _.label++; y = op[1]; op = [0]; continue;
                case 7: op = _.ops.pop(); _.trys.pop(); continue;
                default:
                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
                    if (t[2]) _.ops.pop();
                    _.trys.pop(); continue;
            }
            op = body.call(thisArg, _);
        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
    }
};
Object.defineProperty(exports, "__esModule", { value: true });
var express = require("express");
var valor_1 = require("../entity/valor");
var app_data_source_1 = require("../app-data-source");
var valor = express();
valor.use(express.json());
valor.get("/", function (req, res) {
    return __awaiter(this, void 0, void 0, function () {
        var valor;
        return __generator(this, function (_a) {
            switch (_a.label) {
                case 0: return [4 /*yield*/, app_data_source_1.myDataSource.getRepository(valor_1.Valor).find()];
                case 1:
                    valor = _a.sent();
                    return [2 /*return*/, res.send(valor)];
            }
        });
    });
});
valor.get("/:id", function (req, res) {
    return __awaiter(this, void 0, void 0, function () {
        var results;
        return __generator(this, function (_a) {
            switch (_a.label) {
                case 0: return [4 /*yield*/, app_data_source_1.myDataSource.getRepository(valor_1.Valor).findOneBy({ id: +req.params.id })];
                case 1:
                    if (!!(_a.sent())) return [3 /*break*/, 2];
                    return [2 /*return*/, res.status(404).send("id not found")];
                case 2: return [4 /*yield*/, app_data_source_1.myDataSource.getRepository(valor_1.Valor).findOneBy({
                        id: +req.params.id,
                    })];
                case 3:
                    results = _a.sent();
                    _a.label = 4;
                case 4: return [2 /*return*/, res.send(results)];
            }
        });
    });
});
valor.post("/", function (req, res) {
    return __awaiter(this, void 0, void 0, function () {
        var valor, results;
        return __generator(this, function (_a) {
            switch (_a.label) {
                case 0: return [4 /*yield*/, app_data_source_1.myDataSource.getRepository(valor_1.Valor).create(req.body)];
                case 1:
                    valor = _a.sent();
                    return [4 /*yield*/, app_data_source_1.myDataSource.getRepository(valor_1.Valor).save(valor)];
                case 2:
                    results = _a.sent();
                    return [2 /*return*/, res.status(201).send(results)];
            }
        });
    });
});
valor.put("/:id", function (req, res) {
    return __awaiter(this, void 0, void 0, function () {
        var valor, results;
        return __generator(this, function (_a) {
            switch (_a.label) {
                case 0: return [4 /*yield*/, app_data_source_1.myDataSource.getRepository(valor_1.Valor).findOneBy({ id: +req.params.id })];
                case 1:
                    if (!!(_a.sent())) return [3 /*break*/, 2];
                    return [2 /*return*/, res.status(404).send("id not found")];
                case 2: return [4 /*yield*/, app_data_source_1.myDataSource.getRepository(valor_1.Valor).findOneBy({
                        id: +req.params.id,
                    })];
                case 3:
                    valor = _a.sent();
                    _a.label = 4;
                case 4:
                    app_data_source_1.myDataSource.getRepository(valor_1.Valor).merge(valor, req.body);
                    return [4 /*yield*/, app_data_source_1.myDataSource.getRepository(valor_1.Valor).save(valor)];
                case 5:
                    results = _a.sent();
                    return [2 /*return*/, res.send(results)];
            }
        });
    });
});
valor.delete("/:id", function (req, res) {
    return __awaiter(this, void 0, void 0, function () {
        var results;
        return __generator(this, function (_a) {
            switch (_a.label) {
                case 0: return [4 /*yield*/, app_data_source_1.myDataSource.getRepository(valor_1.Valor).findOneBy({ id: +req.params.id })];
                case 1:
                    if (!!(_a.sent())) return [3 /*break*/, 2];
                    return [2 /*return*/, res.status(404).send("id not found")];
                case 2: return [4 /*yield*/, app_data_source_1.myDataSource.getRepository(valor_1.Valor).delete(req.params.id)];
                case 3:
                    results = _a.sent();
                    _a.label = 4;
                case 4: return [2 /*return*/, res.send(results)];
            }
        });
    });
});
module.exports = valor;