# src/swagger-new.json
{
  "swagger": "2.0",
  "info": {
    "title": "PCTE",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:3000/api",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/utilizador/": {
      "get": {
        "tags": [
          "Utilizador"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "first_name": {
                  "type": "string",
                  "example": "Example"
                },
                "last_name": {
                  "type": "string",
                  "example": "Example"
                },
                "email": {
                  "type": "string",
                  "example": "example@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                },
                "admin": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/utilizador/{id}": {
      "get": {
        "tags": [
          "Utilizador"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "first_name": {
                  "type": "string",
                  "example": "Example"
                },
                "last_name": {
                  "type": "string",
                  "example": "Example"
                },
                "email": {
                  "type": "string",
                  "example": "example@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                },
                "admin": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Utilizador"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "first_name": {
                  "example": "any"
                },
                "last_name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "admin": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "first_name": {
                  "type": "string",
                  "example": "Example"
                },
                "last_name": {
                  "type": "string",
                  "example": "Example"
                },
                "email": {
                  "type": "string",
                  "example": "example@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                },
                "admin": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Utilizador"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Utilizador eliminado"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/valor/": {
      "get": {
        "tags": [
          "Valor"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "idComercializador": {
                  "type": "number",
                  "example": 1
                },
                "idPotencia": {
                  "type": "number",
                  "example": 1
                },
                "idTarifario": {
                  "type": "number",
                  "example": 1
                },
                "valorPotencia": {
                  "type": "number",
                  "example": 0
                },
                "valorEnergia": {
                  "type": "number",
                  "example": 0
                },
                "atualizado": {
                  "type": "number",
                  "example": 1995
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Valor"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "idComercializador": {
                  "type": "number",
                  "example": 1
                },
                "idPotencia": {
                  "type": "number",
                  "example": 1
                },
                "idTarifario": {
                  "type": "number",
                  "example": 1
                },
                "valorPotencia": {
                  "type": "number",
                  "example": 0
                },
                "valorEnergia": {
                  "type": "number",
                  "example": 0
                },
                "atualizado": {
                  "type": "number",
                  "example": 1995
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/valor/{id}": {
      "get": {
        "tags": [
          "Valor"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "idComercializador": {
                  "type": "number",
                  "example": 1
                },
                "idPotencia": {
                  "type": "number",
                  "example": 1
                },
                "idTarifario": {
                  "type": "number",
                  "example": 1
                },
                "valorPotencia": {
                  "type": "number",
                  "example": 0
                },
                "valorEnergia": {
                  "type": "number",
                  "example": 0
                },
                "atualizado": {
                  "type": "number",
                  "example": 1995
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Valor"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "idComercializador": {
                  "type": "number",
                  "example": 1
                },
                "idPotencia": {
                  "type": "number",
                  "example": 1
                },
                "idTarifario": {
                  "type": "number",
                  "example": 1
                },
                "valorPotencia": {
                  "type": "number",
                  "example": 0
                },
                "valorEnergia": {
                  "type": "number",
                  "example": 0
                },
                "atualizado": {
                  "type": "number",
                  "example": 1995
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Valor"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "raw": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "affected": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}